refactor: move augmenters code to augmenters.rs
