+++ src/auth/auth.rs
@@ -1,48 +1,15 @@
 use std::collections::HashMap;
 use std::sync::Arc;
 
+use crate::augmenters::{create_auth_augmenter, Augmenter, AugmenterConfig};
 use crate::config::AuthConfig;
 use crate::models::User;
 use crate::providers::{create_auth_provider, Provider, ProviderConfig};
 use crate::store::Store;
 use futures::future::{join_all, select_ok, FutureExt};
 use futures::lock::Mutex;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
 use tokio::time::timeout;
 use tracing::{debug, info, warn};
-
-use super::ldap_augmenter::{LDAPAugmenter, LDAPAugmenterConfig};
-
-use super::plain_augmenter::{PlainAugmenter, PlainAugmenterConfig};
-
-/// Configuration options for augmenters (e.g. an LDAP roles augmenter).
-#[derive(Deserialize, Serialize, JsonSchema, Debug)]
-#[serde(tag = "type")]
-pub enum AugmenterConfig {
-    #[serde(rename = "ldap")]
-    LDAPAugmenterConfig(LDAPAugmenterConfig),
-
-    #[serde(rename = "plain")]
-    PlainAugmenterConfig(PlainAugmenterConfig),
-}
-
-/// An augmenter can add extra roles or info to an already-authenticated User.
-#[async_trait::async_trait]
-pub trait Augmenter: Send + Sync {
-    fn get_name(&self) -> &str;
-    fn get_type(&self) -> &str;
-    fn get_realm(&self) -> &str;
-    async fn augment(&self, user: Arc<Mutex<User>>) -> Result<(), String>;
-}
-
-/// Create an augmenter from a given config.
-pub fn create_auth_augmenter(config: &AugmenterConfig) -> Box<dyn Augmenter> {
-    match config {
-        AugmenterConfig::LDAPAugmenterConfig(cfg) => Box::new(LDAPAugmenter::new(cfg)),
-        AugmenterConfig::PlainAugmenterConfig(cfg) => Box::new(PlainAugmenter::new(cfg)),
-    }
-}
 
 /// Holds all authentication providers, augmenters, and a reference to the store.
 pub struct Auth {
